# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CHELSY Restaurant API",
        "description": "API REST pour l'application mobile CHELSY Restaurant",
        "contact": {
            "email": "contact@chelsy-restaurant.bj"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/v1/payments/stripe/confirm": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Confirmer un paiement Stripe",
                "operationId": "d8a928d7b0e415b70a1b70373b523ab6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "payment_intent_id"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "payment_intent_id": {
                                        "type": "string",
                                        "example": "pi_xxxxxxxxxxxxx"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement confirmé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/payments/mobile-money/create": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Créer un paiement Mobile Money (FedaPay)",
                "operationId": "b02f30dcf769f804b6f6d30955396fbb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "provider",
                                    "phone_number"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "provider": {
                                        "type": "string",
                                        "enum": [
                                            "MTN",
                                            "Moov"
                                        ],
                                        "example": "MTN"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+229 12 34 56 78"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction Mobile Money créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/payments/mobile-money/status": {
            "post": {
                "tags": [
                    "Paiements"
                ],
                "summary": "Vérifier le statut d'un paiement Mobile Money",
                "operationId": "120102aff7c09840f8c726eb00e29e0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Statut du paiement",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification par token Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}